requirements.txt
fastapi
uvicorn
openfoodfacts
httpx
python-dotenv
groq